+++ nautilus_trader/adapters/polymarket/execution.py
@@ -146,8 +146,11 @@ class PolymarketExecutionClient(LiveExecutionClient):
 
         # Configuration
         self._config = config
+        self._log.info(f"{config.signature_type=}", LogColor.BLUE)
+        self._log.info(f"{config.funder=}", LogColor.BLUE)
         self._log.info(f"{config.max_retries=}", LogColor.BLUE)
         self._log.info(f"{config.retry_delay=}", LogColor.BLUE)
+        self._log.info(f"{config.generate_order_history_from_trades=}", LogColor.BLUE)
 
         account_id = AccountId(f"{name or POLYMARKET_VENUE.value}-001")
         self._set_account_id(account_id)
@@ -161,7 +164,7 @@ class PolymarketExecutionClient(LiveExecutionClient):
         self._log.info(f"{wallet_address=}", LogColor.BLUE)
 
         # HTTP API
-        self._signature_type = 0
+        self._signature_type = config.signature_type
         self._http_client = http_client
         self._retry_manager_pool = RetryManagerPool[None](
             pool_size=100,